requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
cryptography==44.0.0
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
